[add]steamのデータを使って積み上げ面グラフ実装
