+++ src/components/popularity/utils/UnixToDate.ts
@@ -1,7 +1,6 @@
 const UnixToDate = (unix: number):Date => {
 
   const date = new Date(unix * 1000);
-  console.log(date)
 
   return date
 }
