+++ src/hooks/popularity/useCountSteamReviews.ts
@@ -3,7 +3,7 @@ import useSWR from 'swr';
 const fetcher = (url: string) => fetch(url).then((res) => res.json());
 
 const useCountSteamReview = (gameId: number) => {
-  const { data, error } = useSWR(`/api/countSteamReviews`, fetcher);
+  const { data, error } = useSWR(`/api/popularity/countSteamReviews`, fetcher);
   return {
     data,
     error,
